refractor(route): Update customer routes and controllers
