+++ backend/index.ts
@@ -5,6 +5,7 @@ import employeeRouter from "./routes/employee";
 import tipsRouter from "./routes/tips";
 import eventRouter from "./routes/event";
 import statRouter from "./routes/statistics";
+import customersRouter from "./routes/customers";
 
 dotenv.config();
 const app = express();
@@ -15,8 +16,8 @@ const PORT = process.env.PORT || 3000;
 
 const userController = require("./controllers/user");
 const authMiddleware = require("./middleware/auth");
+const managerAuthMiddleware = require("./middleware/managerAuth");
 const fetchController = require("./fetch/fetch");
-const customerController = require("./controllers/customers");
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -25,25 +26,12 @@ app.use("/employee", employeeRouter);
 app.use("/events", eventRouter);
 app.use("/tips", tipsRouter);
 app.use("/statistics", statRouter);
-
+app.use("/customers", customersRouter);
 
 app.post("/login", userController.login);
 
 app.get("/", (request: Request, response: Response) => {
 	response.status(200).send("Hello World");
-});
-
-app.get("/customers", customerController.getCustomers);
-app.get("/customers/:id", customerController.getCustomer);
-app.get("/customers/:id/image", customerController.getCustomerImage);
-app.get("/customers/:id/clothes", customerController.getCustomerClothes);
-app.get("/customers/:id/payments", customerController.getCustomerPayments);
-app.get("/customers/:id/encounters", customerController.getCustomerEncounters);
-
-// Example of a protected route (You can only access this route if you have a valid token)
-// The authMiddleware function checks if the token is valid
-app.get("/protected", authMiddleware, (request: Request, response: Response) => {
-	response.status(200).send("Protected route");
 });
 
 app.listen(PORT, () => {
@@ -55,4 +43,4 @@ app.listen(PORT, () => {
 // Fetch data from API on server start
 fetchController();
 // And then fetch data from API every X minutes (set in a .env file)
-// setInterval(fetchController, parseInt(process.env.FETCH_INTERVAL as string) * 60000);
+setInterval(fetchController, parseInt(process.env.FETCH_INTERVAL as string) * 60000);
